BUG: SparseDataFrame construction with lists not coercing to dtype  (GH 15682)
closes #15682

Author: Carlos Souza <carlos@udacity.com>

Closes #15834 from ucals/bug-fix-15682 and squashes the following commits:

04fba8d [Carlos Souza] Adding test_rename test cases (were missing)
483bb2c [Carlos Souza] Doing adjustments as per @jreback requests
cc4c15b [Carlos Souza] Fixing coersion bug at SparseDataFrame construction
faa5c5c [Carlos Souza] Merge remote-tracking branch 'upstream/master'
43456a5 [Carlos Souza] Merge remote-tracking branch 'upstream/master'
8b463cb [Carlos Souza] Merge remote-tracking branch 'upstream/master'
9fc617b [Carlos Souza] Merge remote-tracking branch 'upstream/master'
e12bca7 [Carlos Souza] Sync fork
676a4e5 [Carlos Souza] Test

+++ pandas/sparse/frame.py
@@ -142,7 +142,7 @@ class SparseDataFrame(DataFrame):
 
         sp_maker = lambda x: SparseArray(x, kind=self._default_kind,
                                          fill_value=self._default_fill_value,
-                                         copy=True)
+                                         copy=True, dtype=dtype)
         sdict = DataFrame()
         for k, v in compat.iteritems(data):
             if isinstance(v, Series):
